requirements.txt
asgiref==3.2.10
Django==3.0.5
django-cleanup==6.0.0
django-crispy-forms==1.9.2
Pillow==9.1.1
pytz==2020.1
sqlparse==0.3.1
